fix : 相互フォローの実装

+++ store/follow.js
@@ -5,6 +5,7 @@ const followRef = db.collection("follow");
 
 export const state = () => ({
   followingUids: [],
+  followedUids: [],
 });
 
 export const actions = {
@@ -22,10 +23,26 @@ export const actions = {
         });
         context.commit("changeFollowingUids", followingUids);
       });
+    followRef
+      .doc(id)
+      .collection("followed")
+      .onSnapshot((snapshot) => {
+        let followedUids = [];
+        snapshot.forEach((doc) => {
+          let id = doc.id;
+          let data = doc.data();
+          data.id = id;
+          followedUids.push(data);
+        });
+        context.commit("changefollowedUids", followedUids);
+      });
   },
   registFollowing(context, items) {
     followRef.doc(items.id).collection("following").add({
       followingUid: items.selectedUid,
+    });
+    followRef.doc(items.selectedUid).collection("followed").add({
+      followedUid: items.id,
     });
   },
   deleteFollowing(context, items) {
@@ -33,23 +50,50 @@ export const actions = {
       .doc(items.id)
       .collection("following")
       .doc(items.deleteItem.id)
-      .delete()
-      .then(() => {
-        // console.log("delete");
+      .delete();
+
+   
+
+    followRef
+      .doc(items.deleteItem.followingUid)
+      .collection("followed")
+      .get()
+      .then((snap) => {
+        snap.docs.forEach((doc) => {
+          if (doc.data().followedUid === items.id) {
+            followRef
+              .doc(items.deleteItem.followingUid)
+              .collection("followed")
+              .doc(doc.id)
+              .delete();
+          } else {
+            return;
+          }
+        });
       });
   },
 };
 
 export const mutations = {
   changeFollowingUids(state, followingUids) {
-    // console.log("change");
-    // console.log(followingUids);
+   
     if (followingUids.length === 0) {
       state.followingUids = [];
     } else {
       state.followingUids = [];
       followingUids.forEach((uid) => {
         state.followingUids.push(uid);
+      });
+    }
+  },
+  changefollowedUids(state, followedUids) {
+   
+    if (followedUids.length === 0) {
+      state.followedUids = [];
+    } else {
+      state.followedUids = [];
+      followedUids.forEach((uid) => {
+        state.followedUids.push(uid);
       });
     }
   },
